+++ CHANGELOG.md
@@ -4,10 +4,19 @@
 
 ### :wrench: Misc
 
-* Update go dependencies:
-  * upgraded github.com/google/go-tpm v0.9.1 => v0.9.3
-  * upgraded golang.org/x/exp v0.0.0-20241210194714-1829a127f884 => v0.0.0-20241217172543-b2144cdd0a67
 * Use os.Root for static file isolation
+* Update go dependencies:
+  * upgraded github.com/c2FmZQ/storage v0.2.3 => v0.2.4
+  * upgraded github.com/c2FmZQ/tpm v0.3.1 => v0.4.0
+  * upgraded github.com/google/go-tpm v0.9.1 => v0.9.3
+  * upgraded github.com/jonboulle/clockwork v0.4.0 => v0.5.0
+  * upgraded github.com/onsi/ginkgo/v2 v2.22.0 => v2.22.2
+  * upgraded golang.org/x/crypto v0.31.0 => v0.32.0
+  * upgraded golang.org/x/exp v0.0.0-20241210194714-1829a127f884 => v0.0.0-20250106191152-7588d65b2ba8
+  * upgraded golang.org/x/net v0.32.0 => v0.34.0
+  * upgraded golang.org/x/sys v0.28.0 => v0.29.0
+  * upgraded golang.org/x/time v0.8.0 => v0.9.0
+  * upgraded golang.org/x/tools v0.28.0 => v0.29.0
 
 ## v0.14.1